+++ public/templates/default/mobile/ajax/lists.php
@@ -15,21 +15,21 @@
                         {if (!$v['answer_info'])}
                         {if $v.is_anonymous}
                         <a href="javascript:;" class="aw-username">
-                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}">
+                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}" loading="lazy" decoding="async">
                         </a>
                         {else/}
                         <a href="{$v['user_info']['url']}" class="aw-user-name" data-id="{$v['user_info']['uid']}">
-                            <img src="{$v['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['user_info']['name']}">
+                            <img src="{$v['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['user_info']['name']}" loading="lazy" decoding="async">
                         </a>
                         {/if}
                         {else/}
                         {if $v['answer_info']['is_anonymous']}
                         <a href="javascript:;" class="aw-username" >
-                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}">
+                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}" loading="lazy" decoding="async">
                         </a>
                         {else/}
                         <a href="{$v['answer_info']['user_info']['url']}" data-id="{$v['answer_info']['uid']}" class="aw-user-name">
-                            <img src="{$v['answer_info']['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['answer_info']['user_info']['name']}">
+                            <img src="{$v['answer_info']['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['answer_info']['user_info']['name']}" loading="lazy" decoding="async">
                         </a>
                         {/if}
                         {/if}
@@ -80,13 +80,13 @@
                     <div class="d-flex aw-list-img">
                         {volist name="$v['img_list']" id="img" key="k"}
                         {if($k<4)}
-                        <img src="{$img|default='/static/common/image/default-cover.svg'}" class="flex-fill mx-1 rounded aw-cut-img" style="margin-bottom: 5px;max-width: 33.3%;border-radius: 5px" >
+                        <img src="{$img|default='/static/common/image/default-cover.svg'}" class="flex-fill mx-1 rounded aw-cut-img" style="margin-bottom: 5px;max-width: 33.3%;border-radius: 5px" loading="lazy" decoding="async">
                         {/if}
                         {/volist}
                     </div>
                     {else/}
                     <div class="aw-list-img" style="height: auto;max-height: 200px;">
-                        <img src="{$v['img_list'][0]|default='/static/common/image/default-cover.svg'}" class="rounded aw-cut-img w-100" style="margin-bottom: 5px;" >
+                        <img src="{$v['img_list'][0]|default='/static/common/image/default-cover.svg'}" class="rounded aw-cut-img w-100" style="margin-bottom: 5px;" loading="lazy" decoding="async">
                     </div>
                      {/if}
                     {/if}
@@ -128,21 +128,21 @@
                         {if (!$v['answer_info'])}
                         {if $v.is_anonymous}
                         <a href="javascript:;" class="aw-username">
-                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}">
+                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}" loading="lazy" decoding="async">
                         </a>
                         {else/}
                         <a href="{$v['user_info']['url']}" class="aw-user-name" data-id="{$v['user_info']['uid']}">
-                            <img src="{$v['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['user_info']['name']}">
+                            <img src="{$v['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['user_info']['name']}" loading="lazy" decoding="async">
                         </a>
                         {/if}
                         {else/}
                         {if $v['answer_info']['is_anonymous']}
                         <a href="javascript:;" class="aw-username" >
-                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}">
+                            <img src="static/common/image/default-avatar.svg" width="40" height="40" class="aw-user-img circle" alt="{:L('匿名用户')}" loading="lazy" decoding="async">
                         </a>
                         {else/}
                         <a href="{$v['answer_info']['user_info']['url']}" data-id="{$v['answer_info']['uid']}" class="aw-user-name">
-                            <img src="{$v['answer_info']['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['answer_info']['user_info']['name']}">
+                            <img src="{$v['answer_info']['user_info']['avatar']}" width="40" height="40" onerror="this.src='static/common/image/default-avatar.svg'" class="circle" alt="{$v['answer_info']['user_info']['name']}" loading="lazy" decoding="async">
                         </a>
                         {/if}
                         {/if}
@@ -193,13 +193,13 @@
                     <div class="d-flex aw-list-img">
                         {volist name="$v['img_list']" id="img" key="k"}
                         {if($k<4)}
-                        <img src="{$img|default='/static/common/image/default-cover.svg'}" class="flex-fill mx-1 rounded aw-cut-img" style="margin-bottom: 5px;max-width: 33.3%;border-radius: 5px" >
+                        <img src="{$img|default='/static/common/image/default-cover.svg'}" class="flex-fill mx-1 rounded aw-cut-img" style="margin-bottom: 5px;max-width: 33.3%;border-radius: 5px" loading="lazy" decoding="async">
                         {/if}
                         {/volist}
                     </div>
                     {else/}
                     <div class="aw-list-img" style="height: auto;max-height: 200px;">
-                        <img src="{$v['img_list'][0]|default='/static/common/image/default-cover.svg'}" class="rounded aw-cut-img w-100" style="margin-bottom: 5px;" >
+                        <img src="{$v['img_list'][0]|default='/static/common/image/default-cover.svg'}" class="rounded aw-cut-img w-100" style="margin-bottom: 5px;" loading="lazy" decoding="async">
                     </div>
                     {/if}
                     {/if}
@@ -239,7 +239,7 @@
                 <div class="aui-card-title pt-3 bg-white px-2 clearfix position-relative pb-2">
                     <div class="float-left">
                         <a href="{$v['user_info']['url']}"  class="aw-user-name" data-id="{$v['user_info']['uid']}" >
-                            <img src="{$v['user_info']['avatar']}" width="40" height="40" onerror="this.src='/static/common/image/default-avatar.svg'" class="circle" alt="{$v['user_info']['name']}">
+                            <img src="{$v['user_info']['avatar']}" width="40" height="40" onerror="this.src='/static/common/image/default-avatar.svg'" class="circle" alt="{$v['user_info']['name']}" loading="lazy" decoding="async">
                         </a>
                     </div>
                     <div class="float-left ml-2">
@@ -274,7 +274,7 @@
                     </div>
                     {if ($v['img_list'] || $v.cover) && get_theme_setting('common.list_show_image')=='Y'}
                     <div class="aw-list-img mb-1" style="max-height: 180px;height: auto">
-                        <img src="{$v.cover|default=$v['img_list'][0]}" class="rounded aw-cut-img" alt="{$v['title']}" width="100%">
+                        <img src="{$v.cover|default=$v['img_list'][0]}" class="rounded aw-cut-img" alt="{$v['title']}" width="100%" loading="lazy" decoding="async">
                     </div>
                     {/if}
                     <div class="aw-content aw-two-line text-muted font-9">